requirements.txt
numpy>=1.24.0
nltk>=3.8.1
tqdm>=4.65.0
PyYAML>=6.0.0
pytest>=7.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
